Rename awsappsync.py to appsync.py
